refactor(butter_bandpass_filter): update imports
