+++ miv/signal/filter/median_filter.py
@@ -9,8 +9,8 @@ import numpy as np
 import numpy.typing as npt
 
 from miv.core.datatype import Signal
-from miv.core.operator_generator import GeneratorOperatorMixin
-from miv.core.wrapper import cache_generator_call
+from miv.core.operator_generator.operator import GeneratorOperatorMixin
+from miv.core.operator_generator.wrapper import cache_generator_call
 from miv.typing import SignalType
