+++ miv/signal/filter/notch_filter.py
@@ -12,8 +12,8 @@ import numpy.typing as npt
 import scipy.signal as sps
 
 from miv.core.datatype import Signal
-from miv.core.operator_generator import GeneratorOperatorMixin
-from miv.core.wrapper import cache_generator_call
+from miv.core.operator_generator.operator import GeneratorOperatorMixin
+from miv.core.operator_generator.wrapper import cache_generator_call
 from miv.typing import SignalType
